feat: add separate key-identifier package
