+++ features/keychain/module/__tests__/eddsa.test.js
@@ -1,6 +1,6 @@
 import { mnemonicToSeed } from 'bip39'
 import { createKeyIdentifierForExodus } from '@exodus/key-ids'
-import { KeyIdentifier } from '../key-identifier'
+import KeyIdentifier from '@exodus/key-identifier'
 import createKeychain from './create-keychain'
 import { getSeedId } from '../crypto/seed-id'
