+++ libraries/key-identifier/jest.config.js
@@ -1,0 +1,5 @@
+/** @type {import('@jest/types').Config.InitialOptions} */
+module.exports = {
+  ...require('../../jest.config.js'),
+  testMatch: ['**/__tests__/**.test.js'],
+}
