+++ libraries/key-identifier/src/index.js
@@ -1,0 +1,1 @@
+export { default } from './key-identifier'
